Refactor import
